Import analyses of leopold2020host experiment
These two analyses of the leopold2020host experiment are imported as
patches from MRM's personal research notebook. The first (earlier)
analysis, 'leopold2020host-bias-sensitivity', conducts an initial
exploration of the sample mean efficiency. The second analysis,
'illustration-of-a-spurious-differential-abundance-result-due-to-bias',
compares the results of regression on calibrated and uncalibrated data
as a case study of the impact of bias on differential abundance.  This
second analysis is the source of the file
'figures/leopold2020host-example.svg'.