Change method of rendering SVGs in PDF output
Previously, I used the cowplot R package to display the SVG images in
the PDF rendering; however, this requires the 'magick' Latex package and
does not always seem to give good results. I now use Inkscape to convert
the SVG to a PDF file in a temporary directory and import this figure
with knitr.