Split original theory out from the manuscript
The original theoretical calculations, which have been put in the
appendix, have largely become redundant and/or superfluous to the
manuscript. In order to improve repo organization and streamline the
rendered manuscript, this commit moves these files to a new theory/
folder and removes them from the rendered manuscript. The theory files
remain structured as a separate Bookdown article.