Revise Solutions section
Besides revising existing sections on calibration, sensitivity analysis,
and meta-analysis, I added a section on ratio-based relative DA analysis
and a section on complementary total-density and sequencing
measurements. I deleted the part of the Implications section on error in
total-density measurement, as this material is now covered by the
'complementary measurements' solution section.