Fix figure rendering in pdf output
The calls to `knitr::include_graphics()` were failing in pdf output
because these calls have to be included in a top level R statement. This
commit fixes this issue; it also removes code redundancy, switches to
relative paths (the recommended method for bookdown), and puts the pdf
images in a new folder contained within the project directoty.