Fix error in leopold2020host analysis
The mean efficiency was calculated incorrectly in the case study
analysis, leading to errors in the plots of mean efficiency and in the
offsets used for bias correction in the linear regression analysis. This
commit fixes the calculation and regenerates the figures and regression
analysis, and updates the main text and supplemental-figure caption to
match the new results.