Add MOMS-PI data
The microbiome data from the MOMS-PI study used in this manuscript's
analysis is contained in Zenodo record
https://zenodo.org/record/7394486, generated from code also contained
there. This commit adds the data to this repo, to make this repository
more self contained. Note, the paths that some of the manuscript Rmd
analysis files use to load this data should be changed to
`here('notebook/_data/momspi')` prior to re-running the analyses.